rearrange files and data
