--- src/test/data/ips.mjs
@@ -1,13 +1,0 @@
-export const goodIPs = [
-  '1.1.1.1',
-  '10.10.10.10',
-  '199.199.199.199',
-  '219.200.200.100',
-  '223.255.255.254'
-]
-
-export const badIPs = [
-  '230.1.1.1',
-  '250.250.250.250',
-  '1.1.1.255'
-]
